moved stuff to folders to be on par with src
