+++ plugin-server/tests/cdp/services/groups-manager.service.test.ts
@@ -1,6 +1,6 @@
-import { GroupsManager } from '../../src/cdp/services/groups-manager.service'
-import { Hub } from '../../src/types'
-import { createHogExecutionGlobals, insertHogFunction as _insertHogFunction } from './fixtures'
+import { GroupsManager } from '../../../src/cdp/services/groups-manager.service'
+import { Hub } from '../../../src/types'
+import { createHogExecutionGlobals, insertHogFunction as _insertHogFunction } from '../fixtures'
 
 describe('Groups Manager', () => {
     jest.setTimeout(1000)
