+++ plugin-server/tests/cdp/services/hog-executor.test.service.ts
@@ -1,14 +1,14 @@
 import { DateTime } from 'luxon'
 
-import { HogExecutor } from '../../src/cdp/services/hog-executor.service'
-import { HogFunctionManager } from '../../src/cdp/services/hog-function-manager.service'
-import { HogFunctionInvocation, HogFunctionType } from '../../src/cdp/types'
-import { Hub } from '../../src/types'
-import { createHub } from '../../src/utils/db/hub'
-import { status } from '../../src/utils/status'
-import { truth } from '../helpers/truth'
-import { HOG_EXAMPLES, HOG_FILTERS_EXAMPLES, HOG_INPUTS_EXAMPLES } from './examples'
-import { createHogExecutionGlobals, createHogFunction, createInvocation } from './fixtures'
+import { HogExecutor } from '../../../src/cdp/services/hog-executor.service'
+import { HogFunctionManager } from '../../../src/cdp/services/hog-function-manager.service'
+import { HogFunctionInvocation, HogFunctionType } from '../../../src/cdp/types'
+import { Hub } from '../../../src/types'
+import { createHub } from '../../../src/utils/db/hub'
+import { status } from '../../../src/utils/status'
+import { truth } from '../../helpers/truth'
+import { HOG_EXAMPLES, HOG_FILTERS_EXAMPLES, HOG_INPUTS_EXAMPLES } from '../examples'
+import { createHogExecutionGlobals, createHogFunction, createInvocation } from '../fixtures'
 
 jest.mock('../../src/utils/status', () => ({
     status: {
