+++ plugin-server/tests/cdp/services/hog-function-manager.service.test.ts
@@ -1,10 +1,10 @@
-import { HogFunctionManager } from '../../src/cdp/services/hog-function-manager.service'
-import { HogFunctionType, IntegrationType } from '../../src/cdp/types'
-import { Hub } from '../../src/types'
-import { closeHub, createHub } from '../../src/utils/db/hub'
-import { PostgresUse } from '../../src/utils/db/postgres'
-import { createTeam, resetTestDatabase } from '../helpers/sql'
-import { insertHogFunction, insertIntegration } from './fixtures'
+import { HogFunctionManager } from '../../../src/cdp/services/hog-function-manager.service'
+import { HogFunctionType, IntegrationType } from '../../../src/cdp/types'
+import { Hub } from '../../../src/types'
+import { closeHub, createHub } from '../../../src/utils/db/hub'
+import { PostgresUse } from '../../../src/utils/db/postgres'
+import { createTeam, resetTestDatabase } from '../../helpers/sql'
+import { insertHogFunction, insertIntegration } from '../fixtures'
 
 describe('HogFunctionManager', () => {
     let hub: Hub
